+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Fixes / Minor Improvements
 
 - Show spinner while updating environment via Collection Explorer.
+- Improve handling for duplicating files via collection explorer:
+    - Open the duplicated file after the process is completed.
+    - For a request file, the duplicated request will be inserted directly after the source file.
 
 ## [1.0.3] (2026-04-27)